requirements.txt
streamlit==1.23.1
Pillow==9.5.0
numpy==1.24.3
opencv-python==4.6.0.66
matplotlib==3.5.3
plotly==5.15.0
tensorflow==2.12.0
keras==2.12.0
